Wavesurfer code partial abstraction
